Fixed RRC/RLC + Compiler warnings
- RRC/RLC were not factoring in the Z flag.

+++ include/timer.h
@@ -32,7 +32,7 @@ class Timer
 
 	private:
 		static int TimerCounter;
-		static BYTE DividerCounter;
+		static int DividerCounter;
 };
 
 #endif
